updates as per review comments
