+++ vehicles/src/modules/special-auth/special-auth.service.ts
@@ -125,4 +125,15 @@ export class SpecialAuthService {
       ReadSpecialAuthDto,
     );
   }
+
+  async findNoFee(companyId: number): Promise<boolean> {
+    const specialAuth = await this.specialAuthRepository
+      .createQueryBuilder()
+      .select()
+      .from(SpecialAuth, 'specialAuth')
+      .innerJoinAndSelect('specialAuth.company', 'company')
+      .where('company.companyId = :companyId', { companyId: companyId })
+      .getOne();
+    return !!specialAuth && !!specialAuth.noFeeType;
+  }
 }
